Import modified code
This work begins the public version of this module.  This work brings a
module that was previously in the shadows for many years, and
contributed to by a few folks on the Puppet Labs Operations team.  This
work is revived here, updated for modern Puppet, adds tests and does a
bit of cleaning.  The base code here has been in use for many years in a
few infrastructures that I am aware of.  Here we just surface that work
to avoid the continuation of copy-paste as we prepare to use this work
in a new infrastructure.